refactor(repository): 删除仓储实现并迁移至模型层
- 移除 `group_repo.go`、`memory_repo.go`、`plan_repo.go` 和 `todo_repo.go` 中旧版仓储实现。
- 删除 `database/repository.go` 和 `database/storm.go`，统一迁移至 GORM 框架。
- 在 `models/entity/` 中新增组实体 `group.go` 和路径映射表支持，优化数据关系。
- 添加 `models/dto/group_dto.go` 以定义组相关的数据传输对象，简化请求数据处理。
- 引入 `models/group_model.go` 实现组管理核心逻辑，包含创建、更新、查找、删除和路径管理功能。
- 重构仓储业务为基于 GORM 的模型架构设计，提升扩展性和性能。
- 替换短连接模式，移除 `OpenWithAction` 和 `OpenWithActionNoReturn` 方法。

+++ internal/tui/models/app.go
@@ -192,11 +192,11 @@ func (m *AppModel) createPage(pageType common.PageType, params map[string]any) c
 		return plan.NewCreateModel(m.bs)
 	case common.PagePlanDetail:
 		id := getIntParam(params, "id")
-		return plan.NewDetailModel(m.bs, id)
+		return plan.NewDetailModel(m.bs, uint(id))
 	case common.PagePlanProgress:
 		id := getIntParam(params, "id")
 		progress := getIntParam(params, "progress")
-		return plan.NewProgressModel(m.bs, id, progress)
+		return plan.NewProgressModel(m.bs, uint(id), progress)
 	case common.PageTodoList:
 		return todo.NewListModel(m.bs)
 	case common.PageTodoToday:
@@ -205,14 +205,14 @@ func (m *AppModel) createPage(pageType common.PageType, params map[string]any) c
 		return todo.NewCreateModel(m.bs)
 	case common.PageTodoDetail:
 		id := getIntParam(params, "id")
-		return todo.NewDetailModel(m.bs, id)
+		return todo.NewDetailModel(m.bs, uint(id))
 	case common.PageGroupList:
 		return group.NewListModel(m.bs)
 	case common.PageGroupCreate:
 		return group.NewCreateModel(m.bs)
 	case common.PageGroupDetail:
 		id := getIntParam(params, "id")
-		return group.NewDetailModel(m.bs, id)
+		return group.NewDetailModel(m.bs, uint(id))
 	default:
 		return NewMenuModel(m.bs)
 	}
